# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "绦管后台 API 文档",
    "description": "电商后台管理系统RESTful API，支持小程序对接",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://qimengzhiyue.cn/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "开发环境"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "用户登录",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          },
          "401": {
            "description": "密码错误"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "用户注册",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功"
          },
          "409": {
            "description": "已存在"
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "获取个人信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "更新个人信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "商品列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "刕efa商品",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          }
        }
      }
    },
...
    "/products/search": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "搜索商品",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "分类列表",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "刕efa分类",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "刕efa订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "items"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "price": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "product_id",
                        "quantity",
                        "price"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          },
          "400": {
            "description": "参数无效"
          },
          "404": {
            "description": "商品不存在"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "用户列表",
        "security": [
          {
            "bearerAuth": [],
            "scopes": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "刕efa用户",
        "security": [
          {
            "bearerAuth": [],
            "scopes": [
              "admin"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "购物车列表",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "添加到购物车",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "productId"
                ],
                "properties": {
                  "productId": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "清空购物车",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    },
...
  }
}